Renamed root element file.
